backend/dnsupdater/requirements-dnsupdater.txt
dnspython==2.1.0
Flask==1.1.2
gunicorn==20.0.4